# Improve performance of eventfilter lookups