@benmoss 2021-04-20
> Am I reading those numbers right? this is a 10x improvement?

Avoiding unnecessary function calls, who knew!