@benmoss 2021-04-20
/test pull-knative-eventing-go-coverage